backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.7.0
langgraph
google-genai
python-multipart
python-dotenv
httpx
weaviate-client

numpy
fpdf2
firebase-admin
